apps/alpr/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
autopep8==2.3.2
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
coloredlogs==15.0.1
dnspython==2.8.0
email-validator==2.3.0
fast-alpr==0.3.0
fast-plate-ocr==1.0.2
fastapi==0.119.1
fastapi-cli==0.0.14
fastapi-cloud-cli==0.3.1
flatbuffers==25.9.23
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
humanfriendly==10.0
idna==3.11
Jinja2==3.1.6
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mpmath==1.3.0
numpy==2.2.6
onnxruntime==1.23.2
open-image-models==0.5.1
opencv-python-headless==4.12.0.88
packaging==25.0
pillow==12.0.0
protobuf==6.33.0
pycodestyle==2.14.0
pydantic==2.12.3
pydantic_core==2.41.4
Pygments==2.19.2
python-dotenv==1.1.1
python-multipart==0.0.20
PyYAML==6.0.3
requests==2.32.5
rich==14.2.0
rich-toolkit==0.15.1
rignore==0.7.1
sentry-sdk==2.42.1
shellingham==1.5.4
sniffio==1.3.1
starlette==0.48.0
sympy==1.14.0
tqdm==4.67.1
typer==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.38.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
